+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Changed (v4.3.0)
+
+## 2019-05-12
+
+- Added new option `selectedAssets` which allows you to pre select Assets when opening the image picker
+
 ### Changed (v4.2.2)
 
 ## 2019-05-08